backend/requirements.txt
fastapi==0.83.0
uvicorn==0.18.3
python-dotenv==0.21.1
supabase==0.7.1
PyJWT==2.4.0
httpx==0.23.3
passlib[bcrypt]==1.7.4
